refactor(api): move progression serializer
