+++ api/tests/evaluation/unit/infrastructure/serializers/jsonapi/progression-serializer_test.js
@@ -1,5 +1,5 @@
-import { expect, domainBuilder } from '../../../../test-helper.js';
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/progression-serializer.js';
+import { expect, domainBuilder } from '../../../../../test-helper.js';
+import * as serializer from '../../../../../../src/evaluation/infrastructure/serializers/jsonapi/progression-serializer.js';
 
 describe('Unit | Serializer | JSONAPI | progression-serializer', function () {
   describe('#serialize()', function () {
